structural changes as discuss in call
