+++ Solutions/IPinfo/Data Connectors/Company/AzureFunctionIPinfoCompany/function.json
@@ -1,0 +1,12 @@
+{
+	"scriptFile": "main.py",
+	"bindings": [
+	  {
+		"name": "myTimer",
+		"type": "timerTrigger",
+		"direction": "in",
+		"schedule": "%SCHEDULE%"
+	  }
+	]
+  }
+
