--- Solutions/IPinfo/Data Connectors/Company/function_app.py
@@ -1,86 +1,0 @@
-import logging
-import maxminddb
-import time
-import azure.functions as func
-from azure.identity import ClientSecretCredential
-from azure.monitor.ingestion import LogsIngestionClient
-from constants import *
-from utils import download_mmdbs
-from utils import check_and_create_data_collection_endpoint
-from utils import check_and_create_table
-from utils import check_and_create_data_collection_rules
-from utils import get_table
-
-app = func.FunctionApp()
-
-@app.schedule(schedule=CRON, arg_name="myTimer", run_on_startup=True, use_monitor=False)
-def ipinfo_company_function(myTimer: func.TimerRequest) -> None:
-    if myTimer.past_due:
-        logging.info("The timer is past due!")
-
-    logging.info("Ipinfo Company timer trigger function executed.")
-
-    def upload_data_to_company_table(dce_endpoint, dcr_immutableid, stream_name):
-        credential = ClientSecretCredential(TENANT_ID, CLIENT_ID, CLIENT_SECRET)
-        client = LogsIngestionClient(endpoint=dce_endpoint, credential=credential, logging_enable=True)
-        mmdb_file_path = "/tmp/standard_company.mmdb"
-        reader = maxminddb.open_database(mmdb_file_path)
-        chunk_size = 10000
-        data_chunk = []
-        logging.info("Uploading Standard Company Data.\n")
-        for ip, ip_data in reader:
-            result = {}
-            result["as_domain"] = ip_data.get("as_domain", "")
-            result["as_name"] = ip_data.get("as_name", "")
-            result["as_type"] = ip_data.get("as_type", "")
-            result["asn"] = ip_data.get("asn", "")
-            result["country"] = ip_data.get("country", "")
-            result["company_domain"] = ip_data.get("domain", "")
-            result["company_name"] = ip_data.get("name", "")
-            result["company_type"] = ip_data.get("type", "")
-            result["range"] = str(ip)
-            data_chunk.append(result)
-            if len(data_chunk) >= chunk_size:
-                try:
-                    client.upload(rule_id=dcr_immutableid, stream_name=stream_name, logs=data_chunk)
-                except Exception as e:
-                    logging.error(f"Upload failed: {e}")
-                data_chunk = []
-        if data_chunk:
-            try:
-                client.upload(rule_id=dcr_immutableid, stream_name=stream_name, logs=data_chunk)
-            except Exception as e:
-                logging.error(f"Upload failed: {e}")
-        reader.close()
-        logging.info("Standard Company Data uploading completed.")
-
-    # Function flow starts here; above this line are function definitions
-    credential = ClientSecretCredential(TENANT_ID, CLIENT_ID, CLIENT_SECRET)
-    access_token = credential.get_token(AZURE_SCOPE).token
-    if access_token:
-        logging.info("\nAccess Token Retrieved\n")
-        logging.info(access_token)
-    else:
-        logging.error("\nFailed to retrieve access token\n")
-
-    download_mmdbs()
-    dce_endpoint = check_and_create_data_collection_endpoint(DATA_COLLECTION_ENDPOINT_NAME, access_token)
-    check_and_create_table(COMPANY_TABLE_NAME, COMPANY_TABLE_SCHEMA, access_token)
-    retries = 3
-    while retries > 0:
-        if get_table(COMPANY_TABLE_NAME, access_token):
-            company_dcr_immutableid, company_stream_name = check_and_create_data_collection_rules(
-                access_token,
-                COMPANY_DCR_NAME,
-                COMPANY_STREAM_DECLARATION,
-                COMPANY_TABLE_COLUMNS,
-                DATA_COLLECTION_ENDPOINT_NAME,
-            )
-            upload_data_to_company_table(dce_endpoint, company_dcr_immutableid, company_stream_name)
-            break
-        else:
-            logging.info("Table not created yet, retrying in 1 minute...")
-            time.sleep(60)
-            retries -= 1
-    if retries == 0:
-        logging.error("Table creation timed out after 3 retries. Data collection rules were not created.")
