+++ Solutions/IPinfo/Data Connectors/Privacy/AzureFunctionIPinfoPrivacy/function.json
@@ -1,0 +1,11 @@
+{
+  "scriptFile": "main.py",
+  "bindings": [
+    {
+      "name": "myTimer",
+      "type": "timerTrigger",
+      "direction": "in",
+      "schedule": "%SCHEDULE%"
+    }
+  ]
+}
