+++ Solutions/IPinfo/Data Connectors/Privacy/azuredeploy_Connector_IPinfo_Privacy_AzureFunction.json
@@ -3,7 +3,7 @@
     "contentVersion": "1.0.0.0",
     "parameters": {
         "FunctionName": {
-            "defaultValue": "IPinfo",
+            "defaultValue": "IPinfo Privacy",
             "minLength": 1,
             "maxLength": 11,
             "type": "string"
@@ -29,6 +29,18 @@
         "IPINFO_TOKEN": {
             "type": "string",
             "defaultValue": "IPinfo Token"
+        },
+        "RETENTION_IN_DAYS": {
+            "type": "string",
+            "defaultValue": "IPinfo Token"
+        },
+        "TOTAL_RETENTION_IN_DAYS": {
+            "type": "string",
+            "defaultValue": "IPinfo Token"
+        },
+        "SCHEDULE": {
+            "type": "string",
+            "defaultValue": "0 30 9 * * *"
         },
         "LOCATION": {
             "type": "string"
@@ -180,6 +192,10 @@
                         "CLIENT_ID": "[parameters('CLIENT_ID')]",
                         "CLIENT_SECRET": "[parameters('CLIENT_SECRET')]",
                         "IPINFO_TOKEN": "[parameters('IPINFO_TOKEN')]",
+                        "RETENTION_IN_DAYS": "[parameters('RETENTION_IN_DAYS')]",
+                        "TOTAL_RETENTION_IN_DAYS": "[parameters('TOTAL_RETENTION_IN_DAYS')]",
+                        "SCHEDULE": "[parameters('SCHEDULE')]",
+                        "LOCATION": "[parameters('LOCATION')]",
                         "WEBSITE_RUN_FROM_PACKAGE": "https://aka.ms/sentinel-IPinfo-Privacy-functionapp"
                     }
                 }
